Move visual studio utils into a dedicated module
